Common: Reorganise files
The common directory was having a lot of files directly under it, which
is not ideal at all as it makes finding a file harder than it should be
so this commit moves some files into subdirectories.
