+++ src/common/auth/auth-checker.js
