+++ src/common/auth/auth-checker.tests.js
