+++ src/common/auth/auth-event-handlers.js
