+++ src/common/auth/session.tests.js
