+++ src/common/components/spinner-component.js
