+++ src/common/components/text-collapse-directive.js
