+++ src/common/forms/form-colorpicker.html
