+++ src/common/index.js
@@ -1,35 +1,35 @@
 import "./footer.css";
 
 import { angular, angularRoute } from "../vendor";
-import Session from "./session";
-import AuthChecker from "./auth-checker";
-import { AUTH_EVENTS, USER_ROLES } from "./auth-constants";
-import authEventHandlers from "./auth-event-handlers";
+import Session from "./auth/session";
+import AuthChecker from "./auth/auth-checker";
+import { AUTH_EVENTS, USER_ROLES } from "./auth/auth-constants";
+import authEventHandlers from "./auth/auth-event-handlers";
 import routeEventHandlers from "./route-event-handlers";
-import AuthInterceptor from "./auth-interceptor";
-import HttpTimeoutInterceptor from "./http-timeout-interceptor";
-import ServerErrorInterceptor from "./server-error-interceptor";
-import FormUploadCtrl from "./form-upload-ctrl";
+import AuthInterceptor from "./auth/auth-interceptor";
+import HttpTimeoutInterceptor from "./http-interceptors/http-timeout-interceptor";
+import ServerErrorInterceptor from "./http-interceptors/server-error-interceptor";
+import FormUploadCtrl from "./forms/form-upload-ctrl";
 import NavigationCtrl from "./navigation-ctrl";
-import SpinnerComponent from "./spinner-component";
-import TextCollapseDirective from "./text-collapse-directive";
-import ShowIfFeatureFlagDirective from "./show-if-feature-flag-directive";
+import SpinnerComponent from "./components/spinner-component";
+import TextCollapseDirective from "./components/text-collapse-directive";
+import ShowIfFeatureFlagDirective from "./components/show-if-feature-flag-directive";
 import configRouting from "./config-routing";
 
 export default angular.module("control.common", [angularRoute, "route-segment", "view-segment"])
-    .service("Session", Session)
-    .service("AuthChecker", AuthChecker)
-    .service("AuthInterceptor", AuthInterceptor)
-    .service("HttpTimeoutInterceptor", HttpTimeoutInterceptor)
-    .service("ServerErrorInterceptor", ServerErrorInterceptor)
-    .constant("AUTH_EVENTS", AUTH_EVENTS)
-    .constant("USER_ROLES", USER_ROLES)
-    .run(authEventHandlers)
-    .run(routeEventHandlers)
-    .controller("FormUploadCtrl", FormUploadCtrl)
-    .controller("NavigationCtrl", NavigationCtrl)
-    .directive("showIfFeatureFlag", ShowIfFeatureFlagDirective)
-    .directive("loadingSpinner", () => SpinnerComponent)
-    .directive("ddTextCollapse", TextCollapseDirective)
-    .config(configRouting)
-    .name;
+  .service("Session", Session)
+  .service("AuthChecker", AuthChecker)
+  .service("AuthInterceptor", AuthInterceptor)
+  .service("HttpTimeoutInterceptor", HttpTimeoutInterceptor)
+  .service("ServerErrorInterceptor", ServerErrorInterceptor)
+  .constant("AUTH_EVENTS", AUTH_EVENTS)
+  .constant("USER_ROLES", USER_ROLES)
+  .run(authEventHandlers)
+  .run(routeEventHandlers)
+  .controller("FormUploadCtrl", FormUploadCtrl)
+  .controller("NavigationCtrl", NavigationCtrl)
+  .directive("showIfFeatureFlag", ShowIfFeatureFlagDirective)
+  .directive("loadingSpinner", () => SpinnerComponent)
+  .directive("ddTextCollapse", TextCollapseDirective)
+  .config(configRouting)
+  .name;
